make cpplint happy

+++ src/test/dag/components.cpp
@@ -18,7 +18,7 @@ TEST_CASE("component complies with the concept", "[dag::component]") {
   STATIC_REQUIRE(dag::network_component<dag::component<int>>);
 
   STATIC_REQUIRE(dag::network_component<
-      dag::component<dag::directed_edge<long int>>>);
+      dag::component<dag::directed_edge<int>>>);
   REQUIRE_NOTHROW(dag::component<int>());
   REQUIRE_NOTHROW(dag::component<
         dag::directed_temporal_hyperedge<std::pair<int, int>, double>>());
